Add missing captions to confirmation pages (#82)
* Create unit tests project

It would be beneficial to begin to create unit tests for complex logic
to prevent regressions when changes are made.

* Create unit tests for `FieldRendererService.RenderFieldAsync`

This method's behaviour is about to change, so asserting its current
behaviour allows for this to be done more intentionally.

* Introduce the current task name to `FieldViewModel`

This will be needed to show the correct caption when confirming the
deletion of an uploaded file.

* Refactor `FieldRendererService` to prevent `NullReferenceException`s

* Update confirmation pages to include the current task where available