Move unit tests project
Currently the GitHub Actions pipelines are failing because the unit
tests are kept in the `Test/` folder. However, this folder isn't copied
into the Docker image, so the solution fails to compile. While this
folder could be added, the Cypress tests are also located in the
`src/Tests` folder, so for consistency, the unit test project has been
moved and the solution file updated.