Fix "back" button redirecting to the same page (#90)
Currently, when using the back button in a task flow, under certain
situations it will redirect the user to the same page instead of
returning to the task summary page.

Specifically, this occurs when the user presses the "Save and continue"
button on the last page in the flow, is taken to the summary page, then
clicks the "Change" link on the last question to go to the last page
again, and then finally clicks the "back" button.

This occurs because the `NavigationHistoryService` is never cleared when
a task flow is completed. Calling the `Clear` method at the appropriate
point in the `RenderFormModel.OnPostPageAsync` view handler prevents
erroneous data from persisting in memory beyond its usefulness and
allows the "back" button to work as expected in this situation.