Move `GenerateSuccessMessage` to `DisplayHelpers` static class
Moving reused pieces of functionality into their own modules reduces the
cognitive load when understanding the rendering logic which makes it
easier to make changes to.

It also allows for testing on this functionality to be done in
isolation, which means that tests can be more specific to the behaviours
and therefore the team can be more confident that changes won't
accidentally introduce new bugs or regressions.

Tests have been added for this method to document its current behaviour.