Add `ExpandEncodedJson` method to `DisplayHelpers`
This acts as a workaround for complex field data being passed around as
a JSON-string-of-JSON rather than as the JSON itself.

To clarify, using adding an academy as an example:

The message "{academiesSearch.name} has been added" expects `itemData`
to have the following shape:
```jsonc
{
  "academiesSearch": {
    "name": "Some Academy",
    "ukprn": "123456"
  },
  // ...
}
```

However, the form engine is providing `itemData` with this shape (note
that `academiesSearch` is a `JsonElement` of kind `String`, rather than
a C# `string`, or a JSON object proper):
```jsonc
{
  "academiesSearch": "{\"name\":\"Some Academy\",\"ukprn\":\"123456\"}",
  // ...
}
```

The `ExpandEncodedJson` method turns `JsonElement` strings like the
latter into `JsonElement` values like the former, while leaving other
values unchanged. After performing this transformation, the
`GenerateSuccessMessage` can perform the correct interpolation on values
given to it by the form engine.