Ensure all field data is available for success message when updating
When a collection item is updated, the user can press the "Change"
button on any of the fields. If they click (for example) the third
button, the values for the first two fields aren't included in
`accumulated`, which results in a bug where success messages show
placeholders instead of the interpolated values. Merging in the original
values using `TryAdd` ensures that all fields are available regardless
of whether they were changed or not.