Allow custom pluralisation when no items are in collection flow
Currently, the `_SingleCollectionFlow.cshtml` view mechanically
pluralises the item label by adding "s" onto the end. In some situations
this is inappropriate: for example, with a task where the
`itemTitleBinding` is "Academy", this results in the incorrect
pluralisation "academys" rather than "academies".

This fixes that problem by allowing tasks in the template schema to
optionally provide a value for `itemTitleBindingPlural`, which if
present will be used instead. When the value is `null` then the view
will fallback to concatenating the item label with "s" as before.