Improve user-facing messages for collection flows (#92)
* Move `GenerateSuccessMessage` to `DisplayHelpers` static class

Moving reused pieces of functionality into their own modules reduces the
cognitive load when understanding the rendering logic which makes it
easier to make changes to.

It also allows for testing on this functionality to be done in
isolation, which means that tests can be more specific to the behaviours
and therefore the team can be more confident that changes won't
accidentally introduce new bugs or regressions.

Tests have been added for this method to document its current behaviour.

* Introduce `JsonElement.EvaluatePath` extension method

This allows `JsonElement`s representing objects to be walked to get a
specific property from them. Using a fully compliant JSON Path
implementation would be better, but `System.Text.Json` doesn't have one
yet and DfE discourages the use of third-party libraries such as
JsonPath.Net.

As this is a greatly simplified implementation, it only supports
dot-based notation, i.e. `foo.bar.baz`. As such, it doesn't allow
addressing of items within arrays.

* Allow custom success messages to interpolate values from `JsonElement`s

Custom success messages can now use the "{key.subkey}" syntax to provide
user-facing messages that take values from complex data fields
represented using `JsonElement`s.

For example, when adding an academy, the success message can be set to
"{academiesSearch.name} has been added" and the `name` property will be
extracted from the complex field named `academiesSearch` and used to
build the final message.

* Add `ExpandEncodedJson` method to `DisplayHelpers`

This acts as a workaround for complex field data being passed around as
a JSON-string-of-JSON rather than as the JSON itself.

To clarify, using adding an academy as an example:

The message "{academiesSearch.name} has been added" expects `itemData`
to have the following shape:
```jsonc
{
  "academiesSearch": {
    "name": "Some Academy",
    "ukprn": "123456"
  },
  // ...
}
```

However, the form engine is providing `itemData` with this shape (note
that `academiesSearch` is a `JsonElement` of kind `String`, rather than
a C# `string`, or a JSON object proper):
```jsonc
{
  "academiesSearch": "{\"name\":\"Some Academy\",\"ukprn\":\"123456\"}",
  // ...
}
```

The `ExpandEncodedJson` method turns `JsonElement` strings like the
latter into `JsonElement` values like the former, while leaving other
values unchanged. After performing this transformation, the
`GenerateSuccessMessage` can perform the correct interpolation on values
given to it by the form engine.

* Ensure all field data is available for success message when updating

When a collection item is updated, the user can press the "Change"
button on any of the fields. If they click (for example) the third
button, the values for the first two fields aren't included in
`accumulated`, which results in a bug where success messages show
placeholders instead of the interpolated values. Merging in the original
values using `TryAdd` ensures that all fields are available regardless
of whether they were changed or not.

* Allow custom pluralisation when no items are in collection flow

Currently, the `_SingleCollectionFlow.cshtml` view mechanically
pluralises the item label by adding "s" onto the end. In some situations
this is inappropriate: for example, with a task where the
`itemTitleBinding` is "Academy", this results in the incorrect
pluralisation "academys" rather than "academies".

This fixes that problem by allowing tasks in the template schema to
optionally provide a value for `itemTitleBindingPlural`, which if
present will be used instead. When the value is `null` then the view
will fallback to concatenating the item label with "s" as before.