Validate radio and checkbox fields receive value in options list
Currently, radio and checkbox fields are not validated (other than to
check if they're set when required), which means that a malicious user
can send any value (e.g. by changing the value in the browser dev
tools). This allows an attacker to execute a cross-site scripting (XSS)
attack.

To prevent this, only the values defined in the template should be
accepted. This updates the `FormValidatorOrchestrator` to fail
validation on radio or checkbox fields where the given value is not one
of the defined options.