Implement HTML sanitisation for user text input
Currently, user input is not sanitised and the user can insert arbitrary
HTML into a page by including it in one of the textarea fields. This
enables a cross-site scripting (XSS) attack as it's possible to inject a
`<script>` tag onto the task summary page and have the browser execute
it.

To prevent this, all user-provided text should be stored in a sanitised
form. The `DisplayHelpers.SanitiseHtmlInput` method uses the built-in
`HtmlEncoder.Encode` method to safely render text.