Introduce display helper to render sanitised text
Now that all user input is sanitised to prevent XSS attacks, there are
some issues when displaying these values:

- Editing a text box or textarea displays the encoded version of the
  original text. For example, if the user entered '5 < 7' and then went
  back to edit that field, it would display '5 &lt; 7'.
- Success messages that interpolate values display the encoded versions
  of those values. For example, if a trustee was added with the name
  'James "Jim" Jones', the success message would read 'James
  &quot;Jim&quot; Jones has been added to Trustees after the transfer'.

In both of these situations, it is desirable to obtain the raw user
input as it was before sanitisation, either because rendering the raw
input is safe (as it is treated as an input value), or because it is
being re-encoded as part of view rendering. For safety reasons, the
application will only ever store the sanitised value, so the best
approach here is to provide a method that acts as the inverse to
`SanitiseHtmlInput`.