Fix cross-site scripting (XSS) vulnerability in template form generation (#96)
* Validate radio and checkbox fields receive value in options list

Currently, radio and checkbox fields are not validated (other than to
check if they're set when required), which means that a malicious user
can send any value (e.g. by changing the value in the browser dev
tools). This allows an attacker to execute a cross-site scripting (XSS)
attack.

To prevent this, only the values defined in the template should be
accepted. This updates the `FormValidatorOrchestrator` to fail
validation on radio or checkbox fields where the given value is not one
of the defined options.

* Implement HTML sanitisation for user text input

Currently, user input is not sanitised and the user can insert arbitrary
HTML into a page by including it in one of the textarea fields. This
enables a cross-site scripting (XSS) attack as it's possible to inject a
`<script>` tag onto the task summary page and have the browser execute
it.

To prevent this, all user-provided text should be stored in a sanitised
form. The `DisplayHelpers.SanitiseHtmlInput` method uses the built-in
`HtmlEncoder.Encode` method to safely render text.

* Introduce display helper to render sanitised text

Now that all user input is sanitised to prevent XSS attacks, there are
some issues when displaying these values:

- Editing a text box or textarea displays the encoded version of the
  original text. For example, if the user entered '5 < 7' and then went
  back to edit that field, it would display '5 &lt; 7'.
- Success messages that interpolate values display the encoded versions
  of those values. For example, if a trustee was added with the name
  'James "Jim" Jones', the success message would read 'James
  &quot;Jim&quot; Jones has been added to Trustees after the transfer'.

In both of these situations, it is desirable to obtain the raw user
input as it was before sanitisation, either because rendering the raw
input is safe (as it is treated as an input value), or because it is
being re-encoded as part of view rendering. For safety reasons, the
application will only ever store the sanitised value, so the best
approach here is to provide a method that acts as the inverse to
`SanitiseHtmlInput`.

* Sanitise form inputs on form page submit

This prevents XSS attacks by storing the user input in an HTML-encoded
format by default.

* Use unsanitised field values in `FieldRendererService`

It is safe to use the unsanitised value to render user input, as it's
either a complex field, in which case it needs to be interpreted as
JSON, or is mapped to an HTML input element, which interprets its value
as text rather than code.

* Unsanitise user-facing messages where double-encoding occurs

In some situations, messages displayed to the user end up being encoded
twice. For example, adding a trustee named 'James "Jim" Jones' ends up
being displayed as 'James &quot;Jim&quot; Jones has been added to
Trustees after the transfer'.

In these situations, it is safe to unsanitise user-input text to ensure
that the messages display correctly, as after doing so it is still
interpreted as text (rather than as raw HTML).