Improve `ItemTitleBinding` handling in form template
Currently, the `ItemTitleBinding` property in a `Task` has two different
but related purposes:
1. Dynamically bind a name to something based on some user-supplied
   value from a field
2. Provide a noun for something

It's not possible to do both at the same time, but it's now necessary to
do so. Therefore, `ItemTitleBinding` has been restricted so that it's
now only used for the first purpose.

To achieve the second purpose, a new `ItemKind` property has been
introduced, and `ItemTitleBindingPlural` has been renamed to
`ItemKindPlural` as it was exclusively used for this second purpose.

Additionally, uses of `ItemTitleBinding` are now processed using
`DisplayHelpers.InterpolateMessage` so that they benefit from the same
`key.subkey` syntax that is already used in success messages.