Escape HTML encoding in `DerivedCollectionFlowService`
Currently, the `DerivedCollectionFlowService` incorrectly reports that
no elements exist in a derived collection flow as now that user input is
sanitised, the JSON used to drive the derived collection cannot be
parsed.

To fix this, the `ProcessCollectionSource` method now HTML decodes the
values of the data dictionaries before attempting to parse the contents
as JSON, restoring the original behaviour.