Fix validation not recognising HTML escaped values for radios
In the declaration, one of the options contains the `'` character, which
gets escaped as `&#27;`, causing the validator to fail as it doesn't
unescape the submitted value before performing the comparison. Using
`HttpUtility.HtmlDecode` on the submitted value provides the correct
behaviour, but causes incorrect rendering on the application review
page, so fields of this type need to be rendered using `@Html.Raw` to
result in a visually correct response.