Dynamically determine names for collection items (#99)
* Fix success message interpolation for complex field data

The fix introduced in #96 to prevent XSS vulnerabilities in
template-driven forms prevents the `DisplayHelpers.ExpandEncodedJson`
method from correctly converting JSON-strings-of-JSON into JSON objects.

To resolve this, the `ExpandEncodedJson` method needs to unsanitise the
contents of JSON strings before attempting to deserialise them.

* Make message placeholder interpolation a public method

This allows the logic to be used outside of generating success messages.

* Improve `ItemTitleBinding` handling in form template

Currently, the `ItemTitleBinding` property in a `Task` has two different
but related purposes:
1. Dynamically bind a name to something based on some user-supplied
   value from a field
2. Provide a noun for something

It's not possible to do both at the same time, but it's now necessary to
do so. Therefore, `ItemTitleBinding` has been restricted so that it's
now only used for the first purpose.

To achieve the second purpose, a new `ItemKind` property has been
introduced, and `ItemTitleBindingPlural` has been renamed to
`ItemKindPlural` as it was exclusively used for this second purpose.

Additionally, uses of `ItemTitleBinding` are now processed using
`DisplayHelpers.InterpolateMessage` so that they benefit from the same
`key.subkey` syntax that is already used in success messages.

* Escape HTML encoding in `DerivedCollectionFlowService`

Currently, the `DerivedCollectionFlowService` incorrectly reports that
no elements exist in a derived collection flow as now that user input is
sanitised, the JSON used to drive the derived collection cannot be
parsed.

To fix this, the `ProcessCollectionSource` method now HTML decodes the
values of the data dictionaries before attempting to parse the contents
as JSON, restoring the original behaviour.

* Fix validation not recognising HTML escaped values for radios

In the declaration, one of the options contains the `'` character, which
gets escaped as `&#27;`, causing the validator to fail as it doesn't
unescape the submitted value before performing the comparison. Using
`HttpUtility.HtmlDecode` on the submitted value provides the correct
behaviour, but causes incorrect rendering on the application review
page, so fields of this type need to be rendered using `@Html.Raw` to
result in a visually correct response.