Setup Cypress to be able to run in GitHub Actions (#101)
* Fix package.json to include correct Cypress dependencies

* Rename Cypress folder to be lowercase

In Unix-like environments (e.g. GitHub Actions), Cypress cannot find its
support files because the folder is currently named `Cypress` (with a
capital C). This works on Windows because paths are case-insensitive
there, but this is not the case for other operating systems.

* Remove unnecessary npm files

Only the `package.json` and `package-lock.json` files in the outer
folder are needed. `tsconfig.json` can also be moved into the outer
folder.

* Setup Cypress to output reports

This has been modelled after https://github.com/DFE-Digital/find-information-about-schools-and-trusts

* Update GitHub Actions workflows to run Cypress tests