Fix duplicate derived collection items when modifying source collection (#116)
A derived collection is dependent on a collection field, e.g.
`trustsSearch-field-flow`. When a user adds a collection item, the
`sourceJson` for the derived collection contains a value with that key.
When a user modifies the collection item, an additional key is added to
the `sourceJson` (in the case of the example above,
`Data[trustsSearch-field-flow]`). This causes the item to be duplicated
in the returned list, so keys of this format need to be ignored.