Add `Caption` property to `Task` model in form templates
This can be used to override the value of the caption in situations
where it should not be the same as the task name.

For example, the "Reason and benefits" task of the "About the trust that
academies are joining" section should use the name of the trust as the
caption (when available) in order to indicate to the user which trust is
being referred to.