Always load accumulated data from session in GET requests
Currently, the `LoadAccumulatedDataFromSessionAsync` method is only
called for tasks that aren't collection flows when a GET is performed.
This causes dynamic captions in collection flows to fail to resolve,
instead displaying their default value.

By moving the method call outside of the conditional that checks whether
or not the current page being requested is part of a collection, this
now always functions as expected.

Additionally, by moving this method call outside of the conditional,
both branches are now identical and the conditional itself can be
removed, simplifying the code.