Allow application templates to define captions for tasks independently of their names (#119)
* Introduce `InterpolatedString` utility object

This represents a string that contains interpolation placeholders within
braces like `{this}`.

It provides an improvement over the existing
`DisplayHelpers.InterpolateMessage` method in that:

- It supports walking through nested `Dictionary`s as well as through
  `JsonElement`s, whereas the previous implementation only supported
  walking through `JsonElement`s
- It supports providing default values when the provided data does not
  have a value for the placeholder. This is done using the
  `{key ?? default value}` syntax.

* Use `InterpolatedString` in `DisplayHelpers.InterpolateMessage`

String interpolation is now handled by `InterpolatedString` so this
method now acts as a forwarding method for backwards compatibility.

* Add `Caption` property to `Task` model in form templates

This can be used to override the value of the caption in situations
where it should not be the same as the task name.

For example, the "Reason and benefits" task of the "About the trust that
academies are joining" section should use the name of the trust as the
caption (when available) in order to indicate to the user which trust is
being referred to.

* Always load accumulated data from session in GET requests

Currently, the `LoadAccumulatedDataFromSessionAsync` method is only
called for tasks that aren't collection flows when a GET is performed.
This causes dynamic captions in collection flows to fail to resolve,
instead displaying their default value.

By moving the method call outside of the conditional that checks whether
or not the current page being requested is part of a collection, this
now always functions as expected.

Additionally, by moving this method call outside of the conditional,
both branches are now identical and the conditional itself can be
removed, simplifying the code.